refactor: extract AnimatedExitRow primitive from SearchList
